Fix nmea test which failed on Arm architecture
The original failure was due to the test sending messages to the node under
test before a connection was established. ROS lacks an API to check if a
connection is established. The first few messages that a publisher sends are
prone to get lost.

The problem is mitigated in this patch by explicitly waiting for the node under
test to reply before moving on with the test.

Issue-Id: SCM-835
Change-Id: Ic521565c989dc9ef0c3d30067d25d5ffc1acb967
Signed-off-by: Liyou Zhou <liyou.zhou@arm.com>