Remove the intermediary pcl::PointCloud structure
Both the input and output are sensor_msg::PointCloud2 and the algorithm
use the pointRH structure. We added a way to directly convert from the
sensor_msg to the pointRH, and then to dissect the input sensor_msg into
two output sensor_msg using the classified data. This is more efficient
than simply converting pointRH to sensor_msg because it allows to skip
the second base change.

Change-Id: I8f92bd01e0b3972f02192443a709166851062f7d
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>