Use an atan approximation instead of the normal one
Copied the atan2 approximation used in autoware.auto into a new .h file.
The approximation is correct at about 0.005 rad in the worst case. This
means a 0.3 degree error which is enough to classify in the wrong ray
with the default parameters. This isn't problematic though for rays have
two purposes : slicing the load into smaller chunk (for the sort) and
approximating the 3d points into a 2d plane. This 0.3 degree of error
create a radius error of around 10^-5 which is still smaller than the
10^-2 of precision the lidar gives us.

Change-Id: Iba32b57af0515e7b7de5b7cf5ebe1190d9d56c97
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>