Remove OpenMP
Omp wasn't actually running in the previous versions, but upon trying
to add it properly, it appears to slows down the execution. Creating
the thread teams (and initializing the variables) takes about 25% as
long as the whole callback, but there is no single part where the
parallelization would gain more time than this overhead would cost.
The only possible candidate would be the transform (~25% of the run
time), but it actually become memory bound instead of CPU bound when
parallelized, voiding the point of parallelizing it.

Change-Id: I0adca8bea85e538688d66873545485ebd569c6c6
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>