Speed up the ray_ground_filter (#12)
* Remove unused code and structure fields

Remove the concentric divider part of ConvertXYZIToRTZColor and shrink
the PointXYZIRTColor structure. In particular, remove the color fields
and the gencolor dependency.

Change-Id: I1850347e5b2e6f83939734edc8b040c996f6c173
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Fix some default parameter comments

Change-Id: I61eab2062ab805a1194af448920124e144987fc5
Issue-ID: SCM-1005
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Move tan call out of the main classify loop

While the variables used in the tan calls are not constant, they are
never changed inside of the main classify loop so we can simply compute
the value once for the whole loop instead of computing it for every
single point

Change-Id: I60470396f5cfb0041eaa59569968d3dbc4a0dbd5
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Use an atan approximation instead of the normal one

Copied the atan2 approximation used in autoware.auto into a new .h file.
The approximation is correct at about 0.005 rad in the worst case. This
means a 0.3 degree error which is enough to classify in the wrong ray
with the default parameters. This isn't problematic though for rays have
two purposes : slicing the load into smaller chunk (for the sort) and
approximating the 3d points into a 2d plane. This 0.3 degree of error
create a radius error of around 10^-5 which is still smaller than the
10^-2 of precision the lidar gives us.

Change-Id: Iba32b57af0515e7b7de5b7cf5ebe1190d9d56c97
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Reserve some memory to reduce the amount of realloc

The constant push_back for every points in every ray was taking a lot
of time (~40% of ConvertXYZIToRTZColor's time). Reserving some space
for those vectors save some time. We can't know each vector's size so
there is still some realloc, but we can at least reserve the average
number of point per vector. We also directly emplace the points
instead of moving them.

Change-Id: I95c692bacee9c3abaa2218b338432b9e24e70e1e
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Shrink even more the internal point structure

We only use the z field in the pcl::pointXYZI to run the algorithm
so we removed the point to only keep this field

Change-Id: Ia19143b037622c76639645b98f1ea8a02c822c6d
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Remove the intermediary pcl::PointCloud structure

Both the input and output are sensor_msg::PointCloud2 and the algorithm
use the pointRH structure. We added a way to directly convert from the
sensor_msg to the pointRH, and then to dissect the input sensor_msg into
two output sensor_msg using the classified data. This is more efficient
than simply converting pointRH to sensor_msg because it allows to skip
the second base change.

Change-Id: I8f92bd01e0b3972f02192443a709166851062f7d
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Add the test for the main callback

Change-Id: Ia6e5b866c27d3be2e1f85ddeead7ff12121547e9
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Remove OpenMP

Omp wasn't actually running in the previous versions, but upon trying
to add it properly, it appears to slows down the execution. Creating
the thread teams (and initializing the variables) takes about 25% as
long as the whole callback, but there is no single part where the
parallelization would gain more time than this overhead would cost.
The only possible candidate would be the transform (~25% of the run
time), but it actually become memory bound instead of CPU bound when
parallelized, voiding the point of parallelizing it.

Change-Id: I0adca8bea85e538688d66873545485ebd569c6c6
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

* Keep all of the original lidar's point

In the oprevious versions, the point above a heigth threshold and
the points too close were filtered out but never added in the
no_ground output. They were simply lost. This patch fix that :
output_ground + output_no_ground = input

Change-Id: Ie2e327871469411c5bd195ea7afe756ed6b1a981
Issue-ID: SCM-1006
Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>

Co-authored-by: Joshua Whitley <josh.whitley@autoware.org>